+++ src/components/List/List.jsx
@@ -1,15 +1,15 @@
 import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { AiFillHeart, AiOutlineHeart, AiOutlineShareAlt } from 'react-icons/ai';
 import { toggleLikeButton } from '../../redux/actions';
 import Grade from '../Grade';
 import Comment from '../Comment';
-import { data } from '../../model/data';
 import { exactPathCopy } from '../../util/index';
 
-const List = ({ dataList }) => {
+const List = () => {
+  const data = useSelector(state => state.data.data);
   const dispatch = useDispatch();
   const [target, setTarget] = useState(null);
   const [isLoaded, setIsLoaded] = useState(false);
@@ -17,6 +17,7 @@ const List = ({ dataList }) => {
   const clickLikeBtn = id => {
     dispatch(toggleLikeButton(id));
   };
+  console.log(data);
 
   const getStarfromRate = reviewRate => {
     let initClicked = [false, false, false, false, false];
@@ -26,43 +27,9 @@ const List = ({ dataList }) => {
     return initClicked;
   };
 
-  const addItems = () => {
-    if (dataList.length === data.length) {
-      alert('더 이상 불러올 데이터가 없습니다.');
-      setIsLoaded(false);
-      return;
-    }
-    setIsLoaded(true);
-    // setDataList(data.slice(0, );
-    setIsLoaded(false);
-  };
-
-  const onIntersect = ([entry], observer) => {
-    if (entry.isIntersecting && !isLoaded) {
-      console.log('check', entry.isIntersecting);
-      observer.unobserve(entry.target);
-      addItems();
-      observer.observe(entry.target);
-    }
-  };
-
-  useEffect(() => {
-    let observer;
-    const defaultOption = {
-      root: null,
-      threshold: 0.5,
-      rootMargin: '0px',
-    };
-    if (target) {
-      observer = new IntersectionObserver(onIntersect, defaultOption);
-      observer.observe(target);
-    }
-    return () => observer && observer.disconnect();
-  }, [target]);
-
   return (
     <Wrapper>
-      {dataList.map(item => {
+      {data.map(item => {
         const {
           id,
           productNm,
