+++ src/redux/actions.js
@@ -1,3 +1,5 @@
+import Filter from "../components/Filter";
+
 export const ADD_COMMENT = 'ADD_COMMENT';
 export const TOGGLE_LIKE = 'TOGGLE_LIKE';
 export const ADD_POST = 'ADD_POST';
@@ -32,23 +34,15 @@ export const addPost = post => {
   };
 };
 
-// {
-//   id: '1',
-//   productNm: '블랙 핸드백',
-//   productImg: [
-//     'https://i.balaan.io/review/c836c897ce27f22497d14d8e9f461ece.webp',
-//   ],
-//   createDt: 1645401600,
-//   review: '무난하게 데일리로 활용중입니다.',
-//   reviewRate: 3,
-//   likeCnt: 244,
-//   comments: [],
-// },
-export const getData = length => {
+export const getData = (data, slice, isLoaded) => {
+  const filetered = data.slice(slice[0], slice[1]);
+  const incomming = [slice[0]+19, slice[1]+19]
   return {
     type: GET_DATA,
     payload: {
-      length,
+      data: filetered,
+      isLoaded,
+      slice: incomming,
     },
   };
 };
