+++ src/redux/reducers/initialState.js
@@ -1,3 +1,6 @@
 import { data } from '../../model/data';
 
-export const initialState = { data };
+export const initialState = {
+  data,
+  slice: [0, 19],
+};
